Limit checking for buying and selling materials from the marketplace

+++ src/include/market.h
@@ -8,6 +8,6 @@
 void fluctuate_market (struct Bot *);
 void print_market (struct Bot *);
 void print_materials (struct Bot *, struct Message *);
-void market_buysell (struct Bot *, struct Message *, int, char *, int);
+void market_buysell (struct Bot *, struct Message *, int, char *, long);
 enum Materials {WOOD, LEATHER, ORE, STONE, BRONZE, MAIL, STEEL, DIAMOND};
 #endif
